refactor: centralize OCI detection and deletion
