+++ lib/modules/manager/kustomize/extract.ts
@@ -7,7 +7,7 @@ import { GitTagsDatasource } from '../../datasource/git-tags';
 import { GithubTagsDatasource } from '../../datasource/github-tags';
 import { HelmDatasource } from '../../datasource/helm';
 import { getDep } from '../dockerfile/extract';
-import { isOCIRegistry } from '../helmv3/utils';
+import { isOCIRegistry, removeOCIPrefix } from '../helmv3/oci';
 import type {
   ExtractConfig,
   PackageDependency,
@@ -149,7 +149,7 @@ export function extractHelmChart(
 
   if (isOCIRegistry(helmChart.repo)) {
     const dep = getDep(
-      `${helmChart.repo.replace('oci://', '')}/${helmChart.name}:${helmChart.version}`,
+      `${removeOCIPrefix(helmChart.repo)}/${helmChart.name}:${helmChart.version}`,
       false,
       aliases,
     );
